+++ src/renderer/components/editor.vue
@@ -67,9 +67,9 @@
 
   const STANDAR_Y = 320
   const PARAGRAPH_CMD = [
-    'ul-bullet', 'ul-task', 'ol-order', 'pre', 'blockquote', 'heading 1', 'heading 2', 'heading 3',
-    'heading 4', 'heading 5', 'heading 6', 'upgrade heading', 'degrade heading', 'paragraph', 'hr',
-    'loose-list-item', 'front-matter'
+    'ul-bullet', 'ul-task', 'ol-order', 'pre', 'blockquote', 'mathblock', 'heading 1', 'heading 2',
+    'heading 3', 'heading 4', 'heading 5', 'heading 6', 'upgrade heading', 'degrade heading',
+    'paragraph', 'hr', 'loose-list-item', 'front-matter'
   ]
 
   export default {
@@ -100,7 +100,7 @@
     },
     computed: {
       ...mapState([
-        'preferLooseListItem', 'autoPairBracket', 'autoPairMarkdownSyntax', 'autoPairQuote'
+        'preferLooseListItem', 'autoPairBracket', 'autoPairMarkdownSyntax', 'autoPairQuote', 'bulletListItemMarker'
       ])
     },
     data () {
@@ -156,12 +156,26 @@
       this.$nextTick(() => {
         const ele = this.$refs.editor
         const {
-          theme, focus: focusMode, markdown, preferLooseListItem, typewriter,
-          autoPairBracket, autoPairMarkdownSyntax, autoPairQuote
+          theme,
+          focus: focusMode,
+          markdown,
+          preferLooseListItem,
+          typewriter,
+          autoPairBracket,
+          autoPairMarkdownSyntax,
+          autoPairQuote,
+          bulletListMarker
         } = this
 
         const { container } = this.editor = new Aganippe(ele, {
-          theme, focusMode, markdown, preferLooseListItem, autoPairBracket, autoPairMarkdownSyntax, autoPairQuote
+          theme,
+          focusMode,
+          markdown,
+          preferLooseListItem,
+          autoPairBracket,
+          autoPairMarkdownSyntax,
+          autoPairQuote,
+          bulletListMarker
         })
 
         if (typewriter) {